complete POC v1
