+++ src/configurations/experiment_poc_2.yaml
@@ -14,7 +14,7 @@ environment:
   game_name: "ALE/Pong-v5"  # Name of the Atari game
   render_mode: "rgb_array"          # Render mode (e.g., "human", "rgb_array", or null)
   max_steps: 10000           # Maximum steps per episode
-  episodes: 100              # Number of training episodes
+  episodes: 500              # Number of training episodes
   batch_size: 32             # Batch size for training
   target_update_frequency: 20  # Update target network every N episodes
   save_options:
