+++ src/main.py
@@ -1,4 +1,3 @@
-import pandas as pd
 import psutil
 import torch
 import time
@@ -6,12 +5,12 @@ import gymnasium as gym
 import ale_py
 import numpy as np
 from src.features.collect import get_exploration_rate
-from src.dqn_agent import DQNAgent
+from src.agents.dqn_agent import DQNAgent
 from src.replay_buffer import ReplayBuffer
-from src.loaders import load_config
+from src.loaders import load_config, save_list_of_dicts_to_dataframe
 
 # load configuration
-config = load_config("src/configurations/experiment_poc.yaml")
+config = load_config("src/configurations/experiment_poc_2.yaml")
 episodes = config["environment"]["episodes"]
 
 # register atari game
@@ -109,6 +108,5 @@ for episode in range(episodes):
     episode_features = {**static_features, **dynamic_features, **memory_features}
     dataset.append(episode_features)
 
-
-df = pd.DataFrame(dataset)
-df.to_csv("dataset_v1.csv")
+# save dataset as dataframe to disk
+save_list_of_dicts_to_dataframe(dataset, config["environment"]["save_options"])
